every button sets its own state
